requirements.txt
discord.py
yt-dlp
ffmpeg
python-dotenv
aiohttp
async-timeout
